\name{freqTab}

\alias{Frequency Table}

\title{Frequency Table}

\description{Create a frequency table of unique values in a data column.}

\usage{freqTab(df,var)}

\arguments{
  \item{df}{this is the dataframe you want to use}
  \item{var}{this is the variable you want to look at the frequency of}
}

\value{A frequency table listing the number of instances each unique value is in a data column, along with the percentage of the total.}

\author{Andrew J. Van Leuven, \email{vanleuven.3@osu.edu}}

\keyword{univar}
