\name{num_of_NAs}

\alias{Number of NAs}

\title{Number of NAs}

\description{Create a table displaying the number of "NA" values in each column.}

\usage{num_of_NAs(x)}

\arguments{
  \item{x}{This is the dataframe you want to use}
}

\value{A table listing the number of "NA" values in each column.}

\author{Andrew J. Van Leuven, \email{vanleuven.3@osu.edu}}

\keyword{univar}
